# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's GIA",
        "description": "API's for transportation management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/gia-backend/public/api/environment": {
            "get": {
                "tags": [
                    "Environment"
                ],
                "summary": "Obtener información con filtros y ordenamiento",
                "operationId": "f41e384d8a3488e9b68367777b12215e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filtrar por descripción",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "route",
                        "in": "query",
                        "description": "Filtrar por ruta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company$business_name",
                        "in": "query",
                        "description": "Filtrar por nombre de empresa",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Entornos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Environment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Environment"
                ],
                "summary": "Crear Environment",
                "operationId": "221a2950959b5fb6c4f63a80b54556c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "company_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del entorno",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descripción del entorno",
                                        "type": "string"
                                    },
                                    "route": {
                                        "description": "Archivo de la imagen",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "Estado del entorno",
                                        "type": "boolean"
                                    },
                                    "company_id": {
                                        "description": "ID de la empresa",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Environment creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Environment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear el environment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/gia-backend/public/api/environment/{id}": {
            "get": {
                "tags": [
                    "Environment"
                ],
                "summary": "Obtener detalles de un environment por ID",
                "operationId": "f537061dfaa60e3fef86e52e8b7f5ae8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Environment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Environment No Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Environment No Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Environment"
                ],
                "summary": "Actualizar Environment por ID",
                "operationId": "2d469c90dae16f08a375bb931c9d4938",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del entorno",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del entorno",
                                        "type": "string",
                                        "example": "Entorno Ejemplo"
                                    },
                                    "description": {
                                        "description": "Descripción del entorno",
                                        "type": "string",
                                        "example": "Descripción del entorno"
                                    },
                                    "route": {
                                        "description": "Archivo de la imagen",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Estado del entorno",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "company_id": {
                                        "description": "ID de la empresa",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Environment actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Environment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Environment No Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Environment No Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al actualizar el environment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Environment"
                ],
                "summary": "Eliminar environment por ID",
                "operationId": "18b82d2975286a6986799d89a37c9ada",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la compañía que se desea eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Environment eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Environment eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Environment No Encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Environment No Encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Company": {
                "title": "Company",
                "description": "Company model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ruc": {
                        "type": "string",
                        "example": "20123456789"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Tech Solutions S.A."
                    },
                    "address": {
                        "type": "string",
                        "example": "Av. Principal 123, Lima"
                    },
                    "phone": {
                        "type": "string",
                        "example": "987654321"
                    },
                    "email": {
                        "type": "string",
                        "example": "info@techsolutions.com"
                    },
                    "route": {
                        "type": "string",
                        "example": "/company/123"
                    }
                },
                "type": "object"
            },
            "Environment": {
                "title": "Environment",
                "description": "Environment model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Production"
                    },
                    "description": {
                        "type": "string",
                        "example": "Primary production environment"
                    },
                    "route": {
                        "type": "string",
                        "example": "/environment/production"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "server_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "company_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}